Reorganize codebase by extracting modules from __main__.py
Extract focused modules to improve organization and eliminate circular
dependency risk between subprocess/worker.py and __main__.py:

- cli.py: CLI types (InputType, UIType, EnumChoice) and validation
- state.py: State management classes (ShrinkRayState and subclasses)
- ui.py: UI abstractions (ShrinkRayUI, BasicUI)
- process.py: Process utilities (signal_group, interrupt_wait_and_kill)
- formatting.py: Formatting utilities (try_decode, default_formatter_command_for, etc.)
- display.py: Display utilities (to_lines, to_blocks, format_diff)

Reduces __main__.py from ~1100 lines to ~355 lines.
